+++ release/nrc/nrc_makah/source/help/nrc_makah.php
@@ -5,14 +5,46 @@
 ?>
 
 <h1>Keyboard Layout</h1>
+<p> This keyboard is for typing the <strong>Makah (Qʷi·qʷi·diččaq)</strong> language.</p>
+<p> <strong>Note</strong>: The mobile layout (for Android, iPhone, iPad)
+operates differently than the desktop layout (for Windows, macOS, Linux). </p>
+
+
 <h2>Mobile/Tablet keyboard layout</h2>
+
+<p>The mobile/tablet is a <strong>longpress</strong> layout. To obtain
+additional letters, <strong>press and hold</strong> the key most similar to
+the desired letter to type it.</p>
+
 <div id="osk-tablet" data-states='default shift numeric'>
 </div>
 
+
+<h2>Desktop keyboard layout</h2>
+<p> The desktop layout implements
+<a href="https://makahmuseum.com/makah-keyboard/">the layout described by the Makah Museum</a>.
+<div id="osk" data-states='default shift rightalt rightalt-shift'>
+</div>
+
+
 <h2> Fonts </h2>
-<p> Fonts that work well for typing Makah on a computer: </p>
+<p> As of this writing (June 7, 2021), many fonts have poor support for Makah
+letters. There are a few letters that can be used to diagnose whether a font
+supports writing in Makah: </p>
+
+<h3> Common problem letters </h3>
+<p>Many fonts struggle with the following characters: </p>
 <ul>
-  <li>Andika</li>
-  <li>Arial</li>
-  <li>Times New Roman</li>
+  <li><strong>ƛ̓</strong> — the glottal mark above often sits too close to ƛ as to be indistiguishable.  </li>
+  <li><strong>x̌</strong> — the &#x25CC;&#x030C; appears off to the side of the x. </li>
+  <li><strong>k̓</strong> — the &#x25CC;&#x0313; appears off to the side of the k. </li>
+  <li><strong>ɫ</strong> — the font may not contain this letter, so a different font is substituted. </li>
+  <li><strong>ʔ</strong> — the font may not contain this letter, so a different font is substituted. </li>
+</ul>
+
+<h3> Fonts that work well for typing Makah on a computer: </h3>
+<ul>
+  <li>Andika: <q style="font-family: Andika">ʔux̌u·ʔaƛa·k,  weʔič̓   šuʔuk̓ʷaƛ̓  ɫ, serifax̌</q></li>
+  <li>Arial: <q style="font-family: Arial">ʔux̌u·ʔaƛa·k,  weʔič̓   šuʔuk̓ʷaƛ̓  ɫ, serifax̌</q></li>
+  <li>Times New Roman: <q style="font-family: Times New Roman">ʔux̌u·ʔaƛa·k,  weʔič̓   šuʔuk̓ʷaƛ̓  ɫ, serifax̌</q></li>
 </ul>