+++ release/nrc/nrc_makah/HISTORY.md
@@ -3,7 +3,8 @@ Makah (Qʷi·qʷi·diččaq) Change History
 
 1.1 (2021-06-07)
 ----------------
-* Add missing image for welcome.htm
+* Add missing image for `welcome.htm`
+* Add desktop layout and font disclaimer to `nrc_makah.php`
 
 1.0 (2021-05-14)
 ----------------